// src/lib/services/common/logger.enum.ts
/***********************************************
********* Copyright mumbler gmbh 2020 **********
************* All rights reserved **************
************************************************/

export enum LogLevel {

	ALL = 0,
	DEBUG = 1,
	VERBOSE = 2,
	WARN = 3,
	ERROR = 4,
	FATAL = 5,
	OFF = 6

}